bot/requirements.txt
beautifulsoup4==4.10.0
discord.py==1.7.3
lxml==4.6.3
matplotlib==3.5.0
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
pytest==6.2.5
dpytest==0.5.3